Pull offerings-related methods into new OfferingFactories file
