+++ common/src/main/java/com/revenuecat/purchases/common/CustomerInfoFactories.kt
@@ -72,57 +72,3 @@ fun JSONObject.buildCustomerInfo(): CustomerInfo {
         originalPurchaseDate = originalPurchaseDate
     )
 }
-
-/**
- * Note: this may return an empty Offerings.
- */
-fun JSONObject.createOfferings(products: Map<String, StoreProduct>): Offerings {
-    val jsonOfferings = getJSONArray("offerings")
-    val currentOfferingID = getString("current_offering_id")
-
-    val offerings = mutableMapOf<String, Offering>()
-    for (i in 0 until jsonOfferings.length()) {
-        jsonOfferings.getJSONObject(i).createOffering(products)?.let {
-            offerings[it.identifier] = it
-
-            if (it.availablePackages.isEmpty()) {
-                warnLog(OfferingStrings.OFFERING_EMPTY.format(it.identifier))
-            }
-        }
-    }
-
-    return Offerings(offerings[currentOfferingID], offerings)
-}
-
-fun JSONObject.createOffering(products: Map<String, StoreProduct>): Offering? {
-    val offeringIdentifier = getString("identifier")
-    val jsonPackages = getJSONArray("packages")
-
-    val availablePackages = mutableListOf<Package>()
-    for (i in 0 until jsonPackages.length()) {
-        jsonPackages.getJSONObject(i).createPackage(products, offeringIdentifier)?.let {
-            availablePackages.add(it)
-        }
-    }
-
-    return if (availablePackages.isNotEmpty()) {
-        Offering(offeringIdentifier, getString("description"), availablePackages)
-    } else {
-        null
-    }
-}
-
-fun JSONObject.createPackage(
-    products: Map<String, StoreProduct>,
-    offeringIdentifier: String
-): Package? {
-    return products[getString("platform_product_identifier")]?.let { product ->
-        val identifier = getString("identifier")
-        val packageType = identifier.toPackageType()
-        return Package(identifier, packageType, product, offeringIdentifier)
-    }
-}
-
-fun String.toPackageType(): PackageType =
-    PackageType.values().firstOrNull { it.identifier == this }
-        ?: if (this.startsWith("\$rc_")) PackageType.UNKNOWN else PackageType.CUSTOM
