+++ common/src/main/java/com/revenuecat/purchases/common/OfferingFactories.kt
@@ -1,0 +1,66 @@
+package com.revenuecat.purchases.common
+
+import com.revenuecat.purchases.Offering
+import com.revenuecat.purchases.Offerings
+import com.revenuecat.purchases.Package
+import com.revenuecat.purchases.PackageType
+import com.revenuecat.purchases.models.StoreProduct
+import com.revenuecat.purchases.strings.OfferingStrings
+import org.json.JSONObject
+
+/**
+ * Note: this may return an empty Offerings.
+ */
+fun JSONObject.createOfferings(productsBySku: Map<String, StoreProduct>): Offerings {
+    val jsonOfferings = getJSONArray("offerings")
+    val currentOfferingID = getString("current_offering_id")
+
+    val offerings = mutableMapOf<String, Offering>()
+    for (i in 0 until jsonOfferings.length()) {
+        val offeringJson = jsonOfferings.getJSONObject(i)
+        offeringJson.createOffering(productsBySku)?.let {
+            offerings[it.identifier] = it
+
+            if (it.availablePackages.isEmpty()) {
+                warnLog(OfferingStrings.OFFERING_EMPTY.format(it.identifier))
+            }
+        }
+    }
+
+    return Offerings(offerings[currentOfferingID], offerings)
+}
+
+fun JSONObject.createOffering(productsBySku: Map<String, StoreProduct>): Offering? {
+    val offeringIdentifier = getString("identifier")
+    val jsonPackages = getJSONArray("packages")
+
+    val availablePackages = mutableListOf<Package>()
+    for (i in 0 until jsonPackages.length()) {
+        val packageJson = jsonPackages.getJSONObject(i)
+        packageJson.createPackage(productsBySku, offeringIdentifier)?.let {
+            availablePackages.add(it)
+        }
+    }
+
+    return if (availablePackages.isNotEmpty()) {
+        Offering(offeringIdentifier, getString("description"), availablePackages)
+    } else {
+        null
+    }
+}
+
+fun JSONObject.createPackage(
+    productsBySku: Map<String, StoreProduct>,
+    offeringIdentifier: String
+): Package? {
+    val sku = getString("platform_product_identifier")
+    return productsBySku[sku]?.let { product ->
+        val identifier = getString("identifier")
+        val packageType = identifier.toPackageType()
+        return Package(identifier, packageType, product, offeringIdentifier)
+    }
+}
+
+fun String.toPackageType(): PackageType =
+    PackageType.values().firstOrNull { it.identifier == this }
+        ?: if (this.startsWith("\$rc_")) PackageType.UNKNOWN else PackageType.CUSTOM
